fix: split topology_builders to satisfy domain test file-size gate
Shard the 709-line world_test_support topology module into focused
sub-files (all under 400 lines) so repo.domain_package_test_file_size
passes on PR #3849.
