+++ packages/colonizethis_world/test/world_test_support/topology_constants.dart
@@ -1,0 +1,10 @@
+import 'package:colonizethis_data/colonizethis_data.dart';
+
+/// Canonical old-world region id for world test support builders.
+const String kWorldTestOw = 'oldWorld';
+
+/// Canonical new-world region id for world test support builders.
+const String kWorldTestNw = 'newWorld';
+
+/// Empty topology for legacy no-op move paths.
+const MapTopology kEmptyMapTopology = MapTopology();
